refactor: clean up unused scripts
Removed some unused `.py` files
