+++ MolecularNodes/pkg.py
@@ -4,9 +4,6 @@ import os
 import site
 from importlib.metadata import version
 import bpy
-import re
-import requests
-import platform
 import pathlib
 
 ADDON_DIR = pathlib.Path(__file__).resolve().parent
@@ -81,3 +78,62 @@ def available():
         except Exception as e:
             all_packages_available = False
     return all_packages_available
+
+class MOL_OT_install_dependencies(bpy.types.Operator):
+    bl_idname = "mol.install_dependencies"
+    bl_label = "Install Dependencies"
+    bl_description = "Install the required python packages to enable import."
+    bl_options = {'REGISTER', 'INTERNAL'}
+    
+    def execute(self, context):
+        if not pkg.available():
+            import datetime,platform
+            
+            # generate logfile
+            logfile_path = os.path.abspath(MolecularNodesAddon.logpath + 'side-packages-install.log')
+            logfile = open(logfile_path, 'a')
+            
+            logfile.write("-----------------------------------" + '\n')
+            logfile.write("Installer Started: " + str(datetime.datetime.now()) + '\n')
+            logfile.write("-----------------------------------" + '\n')
+            install_commands,install_logs=pkg.install(
+                pypi_mirror=bpy.context.scene.pypi_mirror,
+            )
+
+            # log cmd and stdout/stderr
+            for cmd,log in zip(install_commands,install_logs):
+                logfile.write(f'{cmd}\n{log}\n')
+
+                # solve this Apple Silicon installation failure 
+                if ("fatal error: 'Python.h' file not found" in log) and (platform.system()== "Darwin") and ('arm' in platform.machine()):
+                    error_msg = f"ERROR: Could not find the 'Python.h' header file in version of Python bundled with Blender.\n" \
+                            "This is a problem with the Apple Silicon versions of Blender.\n" \
+                            "Please follow the link to the MolecularNodes GitHub page to solve it manually: \n" \
+                            "https://github.com/BradyAJohnston/MolecularNodes/issues/108#issuecomment-1429384983 "
+                    logfile.write(f"{error_msg}\n")
+                    # print this message to Blender script window
+                    print(error_msg)
+
+
+            logfile.write("###################################" + '\n')
+            logfile.write("Installer finished: " + str(datetime.datetime.now()) + '\n')
+            logfile.write("###################################" + '\n')
+            
+            # close the logfile
+            logfile.close()
+
+        if pkg.available():
+            # bpy.context.preferences.addons['MolecularNodesPref'].preferences.packages_available = True
+            self.report(
+                {'INFO'}, 
+                message='Successfully Installed Required Packages'
+                )
+            
+        else:
+            # bpy.context.preferences.addons['MolecularNodesPref'].preferences.packages_available = False
+            self.report(
+                {'ERROR'}, 
+                message=f'Failed to install required packages. \n {error_msg}. \nPlease check log file for details: {logfile_path}'
+                )
+        
+        return {'FINISHED'}
